fix 出品と編集のバグ修正
-ファイル名が日本語だと削除処理の際に文字化けが発生し、エラーとなるためファイル名をgettime関数で取得しファイル名として保存するように修正
-モーダルの表示タイミングやリダイレクトのタイミングで正常に動作していなかったため修正
-編集処理完了時には商品詳細ページへリダイレクトするように修正 flashメッセージは削除
-二重で削除のparamsが送信された際に、2回目には該当のデータがなくエラーとなるため存在していた場合のみ削除処理を行うように修正

+++ app/controllers/products_controller.rb
@@ -70,7 +70,7 @@ end
         if @product.update(product_params)
             update_productimage(@product)
             respond_to do |format|
-                format.html { redirect_to root_path , notice: "更新しました"}
+                format.html { redirect_to product_path }
                 format.json
             end
         else
@@ -145,7 +145,7 @@ end
             i = 0
             while params[:DeletedImageUrls][i].present? do
                 product_image = product.product_images.find_by(image_url: params[:DeletedImageUrls])
-                product_image.destroy
+                product_image.destroy if product_image.present?
                 i += 1
             end
         end
